tests: moved crypto encryption tests into own module
